Adding support to cancel a headless task (#20416)
Summary:
On UWP, when the app is executing a Background task, the OS can request it to cancel the task (for various reasons). If the app does not cancel its background task quickly, the OS might terminate the app.
This change adds support to cancel the headless task which can be used by native code to forward the cancellation request from the OS to the JS code.
Pull Request resolved: https://github.com/facebook/react-native/pull/20416

Differential Revision: D10052080

Pulled By: cpojer

fbshipit-source-id: 2c0322ebb45f7835739f68bdf82a7100d968c516